+++ js/main.js
@@ -1,6 +1,8 @@
+import { openUploadPictureModal } from './form-modal-window.js';
 import { getPictures } from './modal-big-picture.js';
 import { validateHashtags } from './validation-data.js';
 
 getPictures();
+openUploadPictureModal();
 validateHashtags();
