+++ js/validation-data.js
@@ -1,12 +1,7 @@
-import { stopIsEscapeKey } from './util.js';
-import { onModalEscapeKeydown } from './util.js';
 import { isRepeatElement } from './util.js';
 import { checkStringLength } from './util.js';
 
 const formUploadFoto = document.querySelector('.img-upload__form');
-const uploadFotoInput = document.querySelector('.img-upload__input');
-const overlayForForm = document.querySelector('.img-upload__overlay');
-const closeUploadPictureModalButton = document.querySelector('.img-upload__cancel');
 const hashtagInput = document.querySelector('.text__hashtags');
 const commentsInput = document.querySelector('.text__description');
 const HASH_TAG_PATTERN = /^#[a-zа-яё0-9]{1,19}$/i;
@@ -16,7 +11,6 @@ const MAX_LENGTH_COMMENT = 140;
 const pristine = new Pristine(formUploadFoto, {
   classTo: 'form-group',
   errorTextTag: 'div',
-  errorTextClass: 'invalid',
 });
 
 const validateHashtags = () => {
@@ -66,26 +60,7 @@ pristine.addValidator(commentsInput, (value) => checkStringLength(value, MAX_LEN
   1, false
 );
 
-stopIsEscapeKey(hashtagInput);
-stopIsEscapeKey(commentsInput);
-
-const closeUploadPictureModal = () => {
-  overlayForForm.classList.add('hidden');
-  document.querySelector('body').classList.remove('modal-open');
-  uploadFotoInput.value = '';
-};
-
-closeUploadPictureModalButton.addEventListener('click', () => {
-  closeUploadPictureModal();
-});
-
-uploadFotoInput.addEventListener('change', () => {
-  overlayForForm.classList.remove('hidden');
-  document.querySelector('body').classList.add('modal-open');
-  onModalEscapeKeydown(closeUploadPictureModal);
-});
-
-
-export { validateHashtags };
+export { hashtagInput, commentsInput };
+export { pristine, validateHashtags };
